Add generateItems to generate weight and cost
The function generateItems was added to generate the weight and cost
of each item that will be used in the knapsack problem. The function
takes in the cost and weight arrays alongside the number of items and
generates random values for the weight and cost of each item. A
printArray function was also added to print the arrays to test that
the generateItems function works.

The generateItems function allows for random elements to be generated
for usage in the knapsack problem and the printArray function will allow
for the values in every item to be displayed to make sure that the
algorithms works correctly.