Modified cost/weight ratio and mergeSort to float
The costWeightRatio and sortedRatio arrays were changed to floats and
the division of weight and cost was changed to output a float. The
mergeSort function was also modified to sort float values. The print
array function was renamed printIntArray and a new print function was
added printFloatArray that prints floating point arrays.

Modifying the ratios to be floating point values is necessary for the
ratios to be stored correctly. With the arrays holding floating points
an additional print function had to be added for determining the
contents of the float arrays and mergeSort had to be altered so that it
could sort sortedRatio.