Add knapsackSolve to find item selection
The knapsackSolve function takes in the weight and price of each item
alongside the max weight of the knapsack and the total number of items.
This function solves for the items that should be added and prints the
indices that should be selected and the total profit when selecting
those indices.

The knapsackSolve function actually solves the 0-1knapsackProblem and
prints the desired output as the solution to the problem with the input
given as parameters in the function call.